backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
flask-migrate==4.0.5
flask-limiter==3.5.0

# Database
psycopg2-binary==2.9.9
alembic==1.13.1

# File processing
PyPDF2==3.0.1
python-docx==1.1.0

# AI/ML
scikit-learn==1.3.2
spacy==3.7.2
numpy==1.26.2
google-generativeai==0.8.5

# Validation and serialization
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Environment and configuration
python-dotenv==1.0.0

# OAuth and Authentication
authlib==1.2.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Email services
sendgrid==6.10.0

# Production server
gunicorn==21.2.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk[flask]==1.39.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
factory-boy==3.3.0

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Security
cryptography>=40.0.0
bleach==6.1.0
python-magic==0.4.27